Firebase storage init
Python dependencies updated for firebase storage.
Successful call to flask to schedule events.
UI updated to allow for detailed scheduling metrics.
ContentMachine firebase_storage updated to firebase_firestore.
Iterate through buckets after Firebase storage initialization
