+++ ContentApp/src/app/app.component.html
@@ -1,6 +1,8 @@
 <!-- Scripts -->
 <script src="https://accounts.google.com/gsi/client" async defer></script>
 <script src="https://apis.google.com/js/platform.js?onload=onGoogleLibraryLoad" async defer></script>
+<html lang="en">
+<link href="https://fonts.googleapis.com/icon?family=Material+Icons" rel="stylesheet">
 
 <!-- Navigation bar -->
 <p-menubar [model]="items">
