+++ ContentMachine/src/app.py
@@ -3,6 +3,13 @@ from flask_cors import CORS
 import requests
 import appsecrets
 import text_machine as text_machine
+import storage.firebase_storage as firebase_storage
+import os
+import sys
+
+# This code retrieves the current directory path and appends the '../src' directory to the sys.path, allowing access to modules in that directory.
+current_dir = os.path.dirname(os.path.abspath(__file__))
+sys.path.append(os.path.join(current_dir, "../src"))
 
 app = Flask(__name__)
 CORS(app)
@@ -53,14 +60,20 @@ def facebook_callback():
         return jsonify({ 'Authentication failed.', 500 })
 
 @app.route('/api/schedule-text-posts', methods=['POST'])
-def text_to_content():
+async def text_to_content():
     data = request.json
+    print("🚀 ~ file: app.py:58 ~ data:", data)
     userUuid = data['userUuid']
     content = data['content']
     image = data['image']
     frequency = data['frequency']
+
+    if (userUuid is None or content is None or image is None or frequency is None):
+        return jsonify(result=False)
+    else:
+        firebase_storage.downoad_input_prompts('input_prompts', os.path.join('src', 'input_prompts'))
     
-    bool_result = text_machine.run_text_machine(userUuid, content, image, frequency)
+    bool_result = await text_machine.run_text_machine(userUuid, content, image, frequency)
 
     return jsonify(result=bool_result)
