+++ ContentMachine/src/content/ig_content_repo.py
@@ -5,7 +5,7 @@ import meta_graph_api.meta_tokens as meta_tokens
 from domain.endpoint_definitions import make_api_call
 import media.image_creator as image_creator
 from storage.dropbox_storage import DB_FOLDER_REFORMATTED, upload_file_for_sharing_url
-from storage.firebase_storage import firestore_instance, PostingPlatform
+from storage.firebase_firestore import firestore_instance, PostingPlatform
 import json
 
 # This code retrieves the current directory path and appends the '../src' directory to the sys.path, allowing access to modules in that directory.
