+++ ContentMachine/src/content/medium_content_repo.py
@@ -6,7 +6,7 @@ import utility.text_utils as text_utils
 import media.image_creator as image_creator
 import ai.gpt as gpt
 import json
-from storage.firebase_storage import firestore_instance, PostingPlatform
+from storage.firebase_firestore import firestore_instance, PostingPlatform
 
 # This code retrieves the current directory path and appends the '../src' directory to the sys.path, allowing access to modules in that directory.
 current_dir = os.path.dirname(os.path.abspath(__file__))
