+++ ContentMachine/src/meta_graph_api/meta_tokens.py
@@ -4,7 +4,7 @@ from domain.endpoint_definitions import make_api_call
 import requests
 import appsecrets
 import json
-from storage.firebase_storage import firestore_instance, PostingPlatform
+from storage.firebase_firestore import firestore_instance, PostingPlatform
 
 # This code retrieves the current directory path and appends the '../src' directory to the sys.path, allowing access to modules in that directory.
 current_dir = os.path.dirname(os.path.abspath(__file__))
