+++ ContentMachine/src/requirements.txt
@@ -22,3 +22,8 @@ requests-toolbelt==0.10.1
 Flask==2.3.2
 Flask-Cors==3.0.10
 isort==5.12.0
+Flask[async]==2.3.2
+google-cloud-storage==2.0.0
+firebase-admin==6.1.0
+firebase==4.0.1
+
